backend/requirements.txt
Django==4.2.24
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.4.0
pillow==10.4.0
qrcode==7.4.2
channels==4.2.2
channels_redis==4.2.1
redis==6.1.1
service-identity==24.2.0
sqlparse==0.5.3
tomli==2.2.1
Twisted==25.5.0
txaio==23.1.1
typing_extensions==4.13.2
zope.interface==7.2
